% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alias_samples}
\alias{alias_samples}
\title{alias_samples}
\usage{
alias_samples(samples)
}
\value{
tagList of textInput elements
}
\description{
Generates the text input UI elements for aliasing the samples
}
