% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{barplotformatter}
\alias{barplotformatter}
\title{barplotformatter}
\usage{
barplotformatter(dt, vars, df)
}
\arguments{
\item{dt}{a DTdatatable object}

\item{vars}{a character vector of variables}

\item{df}{initial object}
}
\description{
Applies the formatter to each sample column in turn and adds the next
formatting rule to the DT object for all samples
}
