% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{barplotformfun}
\alias{barplotformfun}
\title{barplotformfun}
\usage{
barplotformfun(dt, id, df)
}
\arguments{
\item{dt}{a DT data table}

\item{id}{the name of the column to colour by genotype}

\item{df}{data frame for scale values}
}
\description{
barplotformfun
}
