% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chr_selectizer}
\alias{chr_selectizer}
\title{chr_selectizer}
\usage{
chr_selectizer(chr)
}
\arguments{
\item{chr}{a charater vector of chromosome names}
}
\value{
a shinyWidgets pickerInput UI element that lets the user select
which chromosomes to display
}
\description{
chr_selectizer
}
