% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eff_extractor}
\alias{eff_extractor}
\title{eff_extractor
split multiple EFF strings into separate individual EFF strings
and convert each to a row in a table}
\usage{
eff_extractor(eff)
}
\arguments{
\item{eff}{an EFF string or multiple comma separated EFF strings}
}
\value{
a tibble with an EFF column where each row is a single EFF string
}
\description{
eff_extractor
split multiple EFF strings into separate individual EFF strings
and convert each to a row in a table
}
