% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eff_parser}
\alias{eff_parser}
\title{eff_parser
parsing snpeff strings EFF field
see: http://snpeff.sourceforge.net/SnpEff_manual.html#eff}
\usage{
eff_parser(df)
}
\arguments{
\item{df}{a dataframe with an EFF column containing and EFF string}
}
\value{
a dataframe with columns for
}
\description{
eff_parser
parsing snpeff strings EFF field
see: http://snpeff.sourceforge.net/SnpEff_manual.html#eff
}
