% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/functions.R
\name{effExtractor}
\alias{effExtractor}
\title{effExtractor
split multiple EFF strings into separate individual EFF strings
and convert each to a row in a table}
\usage{
effExtractor(eff)

effExtractor(eff)
}
\arguments{
\item{eff}{an EFF string or multiple comma separated EFF strings}
}
\value{
a tibble with an EFF column where each row is a single EFF string

a tibble with an EFF column where each row is a single EFF string
}
\description{
effExtractor
split multiple EFF strings into separate individual EFF strings
and convert each to a row in a table

effExtractor
split multiple EFF strings into separate individual EFF strings
and convert each to a row in a table
}
