% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/functions.R
\name{effParser}
\alias{effParser}
\title{effParser
parsing snpeff strings EFF field
see: http://snpeff.sourceforge.net/SnpEff_manual.html#eff}
\usage{
effParser(df)

effParser(df)
}
\arguments{
\item{df}{a dataframe with an EFF column containing and EFF string}
}
\value{
a dataframe with columns for

a dataframe with columns for
}
\description{
effParser
parsing snpeff strings EFF field
see: http://snpeff.sourceforge.net/SnpEff_manual.html#eff

effParser
parsing snpeff strings EFF field
see: http://snpeff.sourceforge.net/SnpEff_manual.html#eff
}
