% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gen_del_tab}
\alias{gen_del_tab}
\title{gen_del_tab}
\usage{
gen_del_tab(deltab, samples, aliases)
}
\arguments{
\item{deltab}{data.frame / tibble of loci}

\item{samples}{Vector of sample identifiers}

\item{aliases}{Vector of sample aliases}
}
\description{
Formatting the deletion variant table
}
