% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gen_var_eff_DT}
\alias{gen_var_eff_DT}
\title{gen_var_eff_DT
Generates the datatable of predicted variant effects}
\usage{
gen_var_eff_DT(loci, row_clicked)
}
\arguments{
\item{loci}{a data.frame / tibble of loci}

\item{row_clicked}{the number of a row in the loci table}
}
\value{
a DT data table of the predicted effects of the variant
}
\description{
gen_var_eff_DT
Generates the datatable of predicted variant effects
}
