% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gen_var_tab}
\alias{gen_var_tab}
\title{gen_var_tab}
\usage{
gen_var_tab(loci, samples, aliases)
}
\arguments{
\item{loci}{data.frame / tibble of loci}

\item{samples}{Vector of sample identifiers}

\item{aliases}{Vector of sample aliases}
}
\description{
Formatting the main variant table
}
