% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{genotype_selector}
\alias{genotype_selector}
\title{genotype_selector}
\usage{
genotype_selector(samples, input)
}
\arguments{
\item{samples}{Vector of sample identifiers}

\item{input}{the shiny input object}
}
\value{
UI for set subtraction a tagList of shinyWidgets checkboxGroupButtons
}
\description{
generates genotype filtering UI elements
}
