% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_genotype_class}
\alias{get_genotype_class}
\title{get_genotype_class}
\usage{
get_genotype_class(sig)
}
\arguments{
\item{sig}{genotype signature of the form '0/0'}
}
\value{
a character string with the class of genotype
}
\description{
Classifies genotypes into Homozygous Reference of Alternate alleles or
Heterozygous reference (one allele is reference) and Heterozygous alternate
(both alleles are non-reference and are different from one-another)
}
