% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gtformatter}
\alias{gtformatter}
\title{gtformatter}
\usage{
gtformatter(dt, samples)
}
\arguments{
\item{dt}{a DTdatatable object}

\item{samples}{a character vector of sample names}
}
\value{
a function which take a DT datatable object and fills the genotype
columns of the DT object with different colours depending on the genotype
}
\description{
Applies the formatter to each sample column in turn and adds the next
formatting rule to the DT object for all samples
}
