% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/functions.R
\name{gtformfun}
\alias{gtformfun}
\title{gtformfun}
\usage{
gtformfun(dt, id)

gtformfun(dt, id)
}
\arguments{
\item{dt}{a DT data table}

\item{id}{the name of the column to colour by genotype}
}
\description{
gtformfun

gtformfun
}
