% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_deletions}
\alias{import_deletions}
\title{import_deletions}
\usage{
import_deletions(filenm)
}
\arguments{
\item{filenm}{path to the .gff deletions file generated by MiModD}
}
\value{
a data.frame of deletion variants
}
\description{
import_deletions
}
