% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{layout_ggplotly}
\alias{layout_ggplotly}
\title{layout_ggplotly}
\usage{
layout_ggplotly(gg, x = -0.1, y = -0.04)
}
\arguments{
\item{gg}{and plotly plot object}

\item{x}{a number by which to adjust the vertical position of the x axis name}

\item{y}{a number by which to adjust the horizontal position
of the y axis name}
}
\value{
a plotly plot object
}
\description{
Tweaks the layout of the x and y axis labels so they don't overlap
the axis ticks
from: https://stackoverflow.com/questions/42763280/r-ggplot-and-plotly-axis-margin-wont-change
}
