% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loci_by_genotype}
\alias{loci_by_genotype}
\title{loci_by_genotype}
\usage{
loci_by_genotype(df)
}
\arguments{
\item{df}{a data.frame / tibble for loci}
}
\value{
A tibble with the genotypes spread and the multivariant loci
split into separate rows.
}
\description{
Spreads genotype information into wider format for genotype filtering
splits loci with more than 1 variant into separate rows.
}
