% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mut_type_freq_plot}
\alias{mut_type_freq_plot}
\title{mut_type_freq_plot}
\usage{
mut_type_freq_plot(df)
}
\value{
a plotly plot object
}
\description{
A plot of the number of variant of a given type are currently filtered
}
