% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_geno}
\alias{split_geno}
\title{split_geno}
\usage{
split_geno(dfr)
}
\arguments{
\item{dfr}{a single row from a data.frame / tibble}
}
\value{
A tibble
}
\description{
splits the comma separated fields of multivariate loci, and puts them on
separate rows in a new tibble.
The EFF field requires special treatment as there is not a 1-1 mapping
between number of fields and number of variants, it is split and recombined
appropriately.
}
