% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{variant_column_selector}
\alias{variant_column_selector}
\title{variant_column_selector}
\usage{
variant_column_selector(meta)
}
\arguments{
\item{meta}{data.frame column details (vcfR2tidy metadata table)}
}
\value{
Column selector UI element
}
\description{
generates the column selector UI element for the main variant table
}
