% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wormbase_view}
\alias{wormbase_view}
\title{wormbase_view
Generates an HTML embed tag for a wormbase 'jbrowse' session at the
specified location, optinally with a region highlighted.}
\usage{
wormbase_view(chr, vstart, vend, hstart = NULL, hend = NULL)
}
\arguments{
\item{chr}{the chromosome (no preceeding chr just the numeral)}

\item{vstart}{view starting position}

\item{vend}{view ending position}

\item{hstart}{highlight starting position (optional)}

\item{hend}{highlight ending position (optional)}
}
\value{
character string of html embed tag for jbrowe at locus
}
\description{
wormbase_view
Generates an HTML embed tag for a wormbase 'jbrowse' session at the
specified location, optinally with a region highlighted.
}
